requirements.txt
# Core Testing Framework
behave==1.2.6
selenium==4.15.2
webdriver-manager==4.0.1

# Reporting
allure-behave==2.13.2

# Configuration & API
python-dotenv==1.0.0
requests==2.31.0

# Code Quality Tools
black==23.12.1
flake8==6.1.0
Flake8-pyproject==1.2.3
isort==5.13.2
pylint==3.0.3